refactor(library/Attempt): moves "Success/definition.ts" to "Success/thatYielded.ts"
