+++ src/library/Attempt/Outcome/Success/exports.ts
@@ -1,1 +1,1 @@
-export * from './definition.ts';
+export * from './thatYielded';
