+++ src/library/Attempt/Outcome/Success/thatYielded.ts
